+++ src/main/java/com/robertx22/age_of_exile/saveclasses/perks/PlayerPerksData.java
@@ -8,6 +8,7 @@ import com.robertx22.age_of_exile.database.data.spells.components.Spell;
 import com.robertx22.age_of_exile.database.registry.SlashRegistry;
 import com.robertx22.age_of_exile.saveclasses.PointData;
 import com.robertx22.age_of_exile.uncommon.datasaving.Load;
+import com.robertx22.age_of_exile.uncommon.utilityclasses.LevelUtils;
 import com.robertx22.age_of_exile.vanilla_mc.packets.sync_cap.PlayerCaps;
 import com.robertx22.age_of_exile.vanilla_mc.packets.sync_cap.SyncCapabilityToClient;
 import com.robertx22.library_of_exile.main.Packets;
@@ -21,29 +22,51 @@ import java.util.*;
 public class PlayerPerksData {
 
     @Store
-    public HashMap<String, SchoolData> perks = new HashMap<>();
+    private HashMap<String, SchoolData> perks = new HashMap<>();
+
+    @Store
+    private HashMap<String, SchoolData> spells = new HashMap<>();
 
     @Store
     public int reset_points = 5;
 
-    public int getAllocatedPoints() {
+    public HashMap<String, SchoolData> getPerks(SpellSchool.SchoolType type) {
+        if (type == SpellSchool.SchoolType.SPELLS) {
+            return spells;
+        } else {
+            return perks;
+        }
+    }
+
+    public int getAllocatedPoints(SpellSchool.SchoolType type) {
         int points = 0;
-        for (Map.Entry<String, SchoolData> x : perks.entrySet()) {
+        for (Map.Entry<String, SchoolData> x : getPerks(type).entrySet()) {
             points += x.getValue()
                 .getAllocatedPointsInSchool();
         }
         return points;
     }
 
-    public int getFreePoints(EntityCap.UnitData data) {
-        return (int) ((ModConfig.get().Server.STARTING_TALENT_POINTS + (ModConfig.get().Server.TALENT_POINTS_PER_LVL * data.getLevel())) - getAllocatedPoints());
+    public int getFreePoints(EntityCap.UnitData data, SpellSchool.SchoolType type) {
+
+        int num = 0;
+        if (type == SpellSchool.SchoolType.SPELLS) {
+            num = (int) ModConfig.get().Server.STARTING_SPELL_POINTS;
+            num += ModConfig.get().Server.SPELL_POINTS_AT_MAX_LEVEL / LevelUtils.getMaxLevelMultiplier(data.getLevel());
+
+        } else {
+            num = (int) ModConfig.get().Server.STARTING_TALENT_POINTS;
+            num += ModConfig.get().Server.TALENT_POINTS_AT_MAX_LEVEL / LevelUtils.getMaxLevelMultiplier(data.getLevel());
+        }
+        num -= this.getAllocatedPoints(type);
+        return num;
     }
 
     public SchoolData getSchool(SpellSchool school) {
-        if (!perks.containsKey(school.GUID())) {
-            perks.put(school.GUID(), new SchoolData());
+        if (!getPerks(school.getSchool_type()).containsKey(school.GUID())) {
+            getPerks(school.getSchool_type()).put(school.GUID(), new SchoolData());
         }
-        return perks.get(school.GUID());
+        return getPerks(school.getSchool_type()).get(school.GUID());
     }
 
     public void putOnFirstEmptyHotbarSlot(PlayerEntity player, Spell spell) {
@@ -90,19 +113,23 @@ public class PlayerPerksData {
         getSchool(school).map.put(point, false);
     }
 
-    public boolean hasFreePoints(EntityCap.UnitData data) {
-        return getFreePoints(data) > 0;
+    public boolean hasFreePoints(EntityCap.UnitData data, SpellSchool.SchoolType type) {
+        return getFreePoints(data, type) > 0;
     }
 
     public boolean canAllocate(SpellSchool school, PointData point, EntityCap.UnitData data, PlayerEntity player) {
 
-        if (!hasFreePoints(data)) {
+        if (!hasFreePoints(data, school.getSchool_type())) {
             return false;
         }
 
         Perk perk = school.calcData.getPerk(point);
 
         if (perk.lvl_req > data.getLevel()) {
+            return false;
+        }
+
+        if (perk.isLockedToPlayer(player)) {
             return false;
         }
 
@@ -116,15 +143,11 @@ public class PlayerPerksData {
             }
         }
 
-        if (perk.isLockedToPlayer(player)) {
-            return false;
-        }
-
         return true;
 
     }
 
-    public boolean canRemove(SpellSchool school, PointData point, PlayerEntity player) {
+    public boolean canRemove(SpellSchool school, PointData point) {
 
         if (!getSchool(school).isAllocated(point)) {
             return false;